jsdoc fixes
